renamed SeriesByInit to SeriesAnalysis, removed unused feature_util, updated tests and use case confs to reflect
