+++ internal_tests/pytests/series_analysis/custom.conf
