+++ internal_tests/pytests/series_analysis/series_test.conf
@@ -8,7 +8,7 @@
 LOOP_ORDER = processes
 
 SERIES_ANALYSIS_CONFIG_FILE = {PARM_BASE}/use_cases/feature_relative/met_config/SeriesAnalysisConfig_by_init
-PROCESS_LIST = SeriesByInit
+PROCESS_LIST = SeriesAnalysis
 
 ## NOTE: "TOTAL" is a REQUIRED cnt statistic used by the series analysis scripts
 #STAT_LIST = TOTAL, FBAR, OBAR, ME, MAE, RMSE, BCMSE, E50, EIQR, MAD
