+++ metplus/util/__init__.py
@@ -3,4 +3,3 @@ from .config_metplus import *
 from .time_util import *
 from .met_util import *
 from .string_template_substitution import *
-from .feature_util import *
