+++ metplus/util/met_util.py
@@ -63,8 +63,6 @@ LOWER_TO_WRAPPER_NAME = {'ascii2nc': 'ASCII2NC',
                          'pyembedingest': 'PyEmbedIngest',
                          'regriddataplane': 'RegridDataPlane',
                          'seriesanalysis': 'SeriesAnalysis',
-                         'seriesbyinit': 'SeriesByInit',
-                         'seriesbylead': 'SeriesByLead',
                          'statanalysis': 'StatAnalysis',
                          'tcgen': 'TCGen',
                          'tcpairs': 'TCPairs',
