+++ parm/use_cases/met_tool_wrapper/SeriesAnalysis/SeriesAnalysis.conf
@@ -5,7 +5,7 @@
 [config]
 
 # List of applications to run - only SeriesAnalysis for this case
-PROCESS_LIST = SeriesByInit
+PROCESS_LIST = SeriesAnalysis
 
 # time looping - options are INIT, VALID, RETRO, and REALTIME
 # If set to INIT or RETRO:
