+++ parm/use_cases/model_applications/medium_range/TCStat_SeriesAnalysis_fcstGFS_obsGFS_FeatureRelative_SeriesByLead_PyEmbed_IVT.conf
@@ -11,7 +11,7 @@ TC_PAIRS_CONFIG_FILE = {PARM_BASE}/met_config/TCPairsConfig_wrapped
 
 SERIES_ANALYSIS_CONFIG_FILE = {PARM_BASE}/met_config/SeriesAnalysisConfig_wrapped
 
-PROCESS_LIST = PyEmbedIngest, TCPairs, TCStat, ExtractTiles, SeriesByInit
+PROCESS_LIST = PyEmbedIngest, TCPairs, TCStat, ExtractTiles, SeriesAnalysis
 
 # The init time begin and end times, increment
 # Looping - options are INIT, VALID, RETRO, and REALTIME
