+++ parm/use_cases/model_applications/s2s/GridStat_SeriesAnalysis_fcstNMME_obsCPC_seasonal_forecast.conf
@@ -3,7 +3,7 @@
 [config]
 
 # List of applications to run
-PROCESS_LIST = GridStat, SeriesByInit
+PROCESS_LIST = GridStat, SeriesAnalysis
 
 # loop by INIT time (options are INIT, VALID, or ?)
 LOOP_BY = INIT
